Split ChildrenPlain option into dedicated story file
